feat(analytics): track citation clicks
Extract the citation anchor into a client component emitting citation_click;
Prose re-exports A so the rest stays server-rendered.

Refs #400
